BACKLOG-20304 Extract dnd into separate element, fix move from list to list, multiple updates
