+++ src/javascript/SelectorTypes/MultipleLeftRightSelector/MultipleLeftRightSlector.scss
@@ -15,9 +15,3 @@
     flex-direction: column;
     gap: 10px;
 }
-
-.listHolder {
-    display: flex;
-    flex-grow: 1;
-    flex-direction: column;
-}
